requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=pyhd3eb1b0_0
asttokens=2.2.1=pyhd8ed1ab_0
av=10.0.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py311h5bb9823_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py311h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
certifi=2023.5.7=py311haa95532_0
cffi=1.15.1=py311h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.5=py311h59b6b97_0
cryptography=39.0.1=py311h21b164f_0
cuda-cccl=12.2.53=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py311hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.9.0=py311haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py311haa95532_0
imageio=2.26.0=py311haa95532_0
imageio-ffmpeg=0.4.8=pypi_0
importlib-metadata=6.7.0=pyha770c72_0
importlib_metadata=6.7.0=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.15.0=pyh025b116_0
ipython=8.14.0=pyh08f2357_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py311haa95532_0
jpeg=9e=h2bbff1b_1
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h1ea47a8_0
kiwisolver=1.4.4=py311hd77b12b_0
krb5=1.19.4=h5b6d351_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.3.53=0
libcurand-dev=10.3.3.53=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py311h2bbff1b_0
matplotlib=3.7.1=py311haa95532_1
matplotlib-base=3.7.1=py311hf62ec03_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.6=py311hf62ec03_1
mkl_random=1.2.2=py311hf62ec03_1
mpmath=1.2.1=py311haa95532_0
munkres=1.1.4=py_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=2.8.4=py311haa95532_1
numexpr=2.8.4=py311h1fcbade_1
numpy=1.25.0=py311hdab7c0b_0
numpy-base=1.25.0=py311hd01c5d8_0
openssl=1.1.1u=h2bbff1b_0
packaging=23.0=py311haa95532_0
pandas=1.5.3=py311heda8569_0
parso=0.8.3=pyhd8ed1ab_0
pathlib=1.0.1=pyhd3eb1b0_1
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311hd77b12b_0
pip=23.1.2=py311haa95532_0
platformdirs=2.5.2=py311haa95532_0
ply=3.11=py311haa95532_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.0=py311h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=23.0.0=py311haa95532_0
pyparsing=3.0.9=py311haa95532_0
pyqt=5.15.7=py311hd77b12b_0
pyqt5-sip=12.11.0=py311hd77b12b_0
pysocks=1.7.1=py311haa95532_0
python=3.11.3=h966fe2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.11=2_cp311
pytorch=2.0.1=py3.11_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-mutex=1.0=cuda
pytz=2022.7=py311haa95532_0
pywin32=305=py311h2bbff1b_0
pyzmq=25.1.0=py311hd77b12b_0
qt-main=5.15.2=he8e5bd7_8
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h2bbfb41_5
requests=2.29.0=py311haa95532_0
scipy=1.10.1=py311hc1ccb85_1
setuptools=67.8.0=py311haa95532_0
sip=6.6.2=py311hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py311haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=2.0.2=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.2=py311h2bbff1b_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.6.3=py311haa95532_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.2=h8cc25b3_0
zeromq=4.3.4=h0e60522_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
